AlignTune/requirements.txt
# Core
accelerate>=0.30.0
datasets>=2.19.0
evaluate>=0.4.2
huggingface-hub>=0.23.0
transformers>=4.42.0
peft>=0.11.0
trl>=0.9.0
bitsandbytes; platform_system != "Windows"
scikit-learn>=1.3.0
numpy>=1.24.0
rouge-score>=0.1.2
pyyaml>=6.0.0

# Serving & UI
uvicorn>=0.24.0
fastapi>=0.110.0
gradio>=4.0.0

# Dev
pytest>=7.4.0
flake8>=6.1.0

# Optional accelerators / serving (install if you need them)
# unsloth>=2024.6.0
# flash-attn>=2.5.8
# vllm>=0.5.0
